backend/pyproject.toml
[project]
name = "arex-backend"
version = "0.1.0"
description = "AREX Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "google-cloud-storage>=2.19.0",
    "python-multipart>=0.0.20",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.44.0",
]

[tool.uv]
package = false
